Update C++ base class
This commit tries to make MiDispositivoMIDI_V3 more flexible by
detaching the logic that reads the PADs from the function they
produce.
Now readPADs produces a PADstatus {OFF, ON, ONSET, OFFSET} that is
passed to PADaction.
PADaction sends MIDI notes by default, but it is defined as a virtual
function so the user can inherit it and define custom behaviors for
the PADs.