Add prototype of Automelody example (WIP)
This example generates melodies following a probalistic model
generated from MIDI files. Note probabilities and transition
probabilities are considered in the decision of the next note.
PADs 1 to 8 will control the note generation (things like going
up or down in the scale)
PADs 9 to 16 will reserved for controling which scale to use,
octave up/down and maybe to add some randomness on the note selection...
Still to be decided!