Add support for a 4x4 potentiometer extension
Add support for a 4x4 potentiometers extension. With this new feature you can have your main 4x4 board with touch pads, and up to 3 potentiometers or touch pads extensions. All combinations are supported, 4x16, 8x8, 4x8, and you can choose for each one if it contains a potentiometer board or a toch pad board. Note that the potentiometer board is not sold at the moment, but you can create it with a mux and 16 potentiometers. See doc.